[day1]: Solving fun stuff.
Using `trie` to solve the first day's puzzle.

First part is very simple as we only need to grab an integer.

Second part is a little bit more nuanced since you need to
convert some of the strings into [1-9] in string format.
This can be annoying and longer to code, but not impossible.
Using the `trie` like structure to solve this and have a great runtime.
Even pretty fast on Typescript!

Anyway, pretty cool snippets are line:77 of the day1. Pretty tricky to
configure, but using `i (index of string) and j (index of checking
stringified digits) - 1`, we could traverse back to the original string
index since it will be incremented (that's why it's `-1`).

Thought this would be easy, but a lil hard. :O

3/5 fun, but don't want to repeat.

=== unrelated to AoC

* Removing strict restriction of imports/exports; can't compile
* Implemented `utils.ts` file for reading and parsing.