solve(day4): I f-ing hate split.
Apparently, split needs to be regex'd if you want consistent behavior
across the language. Using ` ` instead of `\s+` for regex, caused
typescript to render something `  1` into undefined or something else of
different size of an array that is unexpected.

Feels like this behavior should be fixed or at least notified by the
program, but apparently, this is all pipe-dream.

Honestly, should not have taken this longer, but it took me about
45minutes due to debugging >:(

I hate `splits`, just split by the white space :'(

Other than snafu with whitespaces in `split()`, it was a fun puzzle.